Get data2mem working
To allow data2mem replacing the memory content the RAM DWORDs must be
ordered low byte first in the bitfile. But this bit order of a bit is 7
downto 0. For this reason byte swaping in module MainMemory.vhd is
implemented. It can be switched of for simulation because it is more
comfortable to work without byte swapping in simulation